requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
psycopg2>=2.7,<3.0